requirements.txt
# Core web framework
flask==2.2.3
gunicorn==20.1.0
Werkzeug==2.2.3

# Database ORM
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.5
psycopg2-binary==2.9.5
alembic==1.10.2

# Environment and configuration
python-dotenv==1.0.0

# API interaction
requests==2.28.2

# File upload handling
python-magic==0.4.27

# MIME type detection
python-magic-bin==0.4.14; sys_platform == 'win32'

# JSON handling
marshmallow==3.19.0
Flask-Marshmallow==0.15.0

# Security
Flask-WTF==1.1.1
email-validator==1.3.1

# Email sending
Flask-Mail==0.9.1

# Background tasks (for production)
celery==5.2.7
redis==4.5.1

# Testing
pytest==7.2.2
pytest-cov==4.0.0

# Development tools
black==23.1.0
flake8==6.0.0
isort==5.12.0